requirements.txt
alembic==1.0.10
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.7.1
betamax==0.8.1
betamax-serializers==0.2.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
docutils==0.14
dotenv==0.0.5
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.17
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
lazy-object-proxy==1.3.1
lockfile==0.12.2
luigi==2.8.3
lxml==4.3.3
Mako==1.0.9
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
numpy==1.16.3
pandas==0.24.2
pluggy==0.9.0
py==1.8.0
pycodestyle==2.5.0
pylint==2.3.1
pytest==4.4.1
pytest-mock==1.10.4
python-daemon==2.1.2
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2019.1
PyYAML==5.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
spotipy==2.4.4
SQLAlchemy==1.3.3
tornado==4.5.3
typed-ast==1.3.4
urllib3==1.24.2
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.2.1
